--- src/app/interfaces/pvt.ts
@@ -1,10 +1,0 @@
-export interface Pvt {
-  type: 'pvt';
-  id: string;
-  trials: number;
-  min_waiting: number;
-  max_waiting: number;
-  show: boolean;
-  max_reaction: number;
-  exit: boolean;
-}
