+++ src/app/interfaces/study.ts
@@ -1,6 +1,3 @@
-import { Pvt } from './pvt';
-import Survey from './survey';
-
 export interface Study {
   properties: Properties;
   modules: Module[];
@@ -58,4 +55,124 @@ export interface Graph {
 export interface Time {
   hours: number;
   minutes: number;
+}
+
+export interface Pvt {
+  type: 'pvt';
+  id: string;
+  trials: number;
+  min_waiting: number;
+  max_waiting: number;
+  show: boolean;
+  max_reaction: number;
+  exit: boolean;
+}
+
+export interface Survey {
+  type: 'survey';
+  id: string;
+  name: string;
+  shuffle: boolean;
+  sections: Section[];
+  submit_text: string;
+}
+
+export interface Section {
+  id: string;
+  name: string;
+  shuffle: boolean;
+  questions: Question[];
+}
+
+export interface Question {
+  id: string;
+  text: string;
+  body:
+    | Instruction
+    | DateTime
+    | Multi
+    | Text
+    | Slider
+    | Media
+    | YesNo
+    | External
+    | File;
+  required: boolean;
+  hide_id: string;
+  hide_value: string;
+  hide_if: boolean;
+  rand_group?: string;
+
+  hidden?: boolean;
+  hideError?: boolean;
+  model?: any;
+  noToggle?: boolean;
+  response?: any;
+  value?: any;
+}
+
+export interface Instruction {
+  type: 'instruction';
+}
+
+export interface YesNo {
+  type: 'yesno';
+  yes_text: string;
+  no_text: string;
+}
+
+export interface Text {
+  type: 'text';
+  subtype: 'short' | 'long' | 'numeric';
+}
+
+export interface DateTime {
+  type: 'datetime';
+  subtype: 'date' | 'time' | 'datetime';
+}
+
+export interface Slider {
+  type: 'slider';
+  min: number;
+  max: number;
+  hint_left: string;
+  hint_right: string;
+}
+
+export interface Multi {
+  type: 'multi';
+  radio: boolean;
+  modal: boolean;
+  options: [];
+  shuffle: boolean;
+
+  optionsChecked?: Option[];
+}
+
+export interface File {
+  type: 'file';
+  src: string;
+  file_name: string;
+}
+
+export interface Media {
+  type: 'media';
+  subtype: 'image' | 'video' | 'audio';
+  src: string;
+  thumb: string;
+
+  safethumb?: any;
+  safeurl?: any;
+}
+
+export interface External {
+  type: 'external';
+  src: string;
+
+  safeurl?: any;
+}
+
+export interface Option {
+  text: string;
+  checked: boolean;
 }
